requirements.txt
openai==0.28.0
requests
python-dotenv
